pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "raad-video"
description = "High-performance video loading library for machine learning"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["video", "machine learning", "data loading", "deep learning"]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "redis>=4.5.0",
    "pyzmq>=24.0.0",
    "requests>=2.26.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
]

[tool.setuptools]
packages = ["raad"]

[tool.setuptools_scm]
write_to = "raad/_version.py"
